Fix for the awardCredits bug.
git-svn-id: http://svn.berlios.de/svnroot/repos/oolite-linux/trunk@1058 127b21dd-08f5-0310-b4b7-95ae10353056

+++ src/Core/Entities/PlayerEntityLegacyScriptEngine.m
@@ -885,7 +885,7 @@ static int scriptRandomSeed = -1;	// ensure proper random function
 	if (script_target != self)  return;
 	
 	int award = 10 * [valueString intValue];
-	if (credits < -award)  credits = 0;
+	if ((int)credits < -award)  credits = 0;
 	else  credits += award;
 }
